By dividing Two numbers

+++ Find_Remainder.py
@@ -1,4 +1,4 @@
-num1=int(input("Enter Number1:\n"))
-num2=int(input("Enter Number2:\n"))
+num1=float(input("Enter Number1:\n"))
+num2=float(input("Enter Number2:\n"))
 rem=num1%num2
 print(f"when {num1} divided by {num2} then remiander is {rem}")
